fix up package rename error
