+++ lib/effects-sqs/src/main/scala/com/gu/effects/sqs/AWSAsyncHandler.scala
@@ -1,4 +1,4 @@
-package manualTest
+package com.gu.effects.sqs
 
 import java.util.concurrent.{Future => JFuture}
