+++ lib/effects-sqs/src/test/scala/com/gu/effects/sqs/AWSSQSSendTest.scala
@@ -1,9 +1,9 @@
-package manualTest
+package com.gu.effects.sqs
 
 import com.amazonaws.regions.Regions
 import com.amazonaws.services.sqs.AmazonSQSAsyncClientBuilder
 import com.amazonaws.services.sqs.model.ReceiveMessageRequest
-import AwsSQSSend.{Payload, QueueName}
+import com.gu.effects.sqs.AwsSQSSend.{Payload, QueueName}
 import com.gu.test.EffectsTest
 import org.apache.log4j.Logger
 import org.scalatest.{AsyncFlatSpec, Matchers}
